Extracted matchers to util
